Refactor weather to mimic Room and Font structure
- Consolidated WeatherSystem and WeatherRenderer into GLWeather in src/opengl/
- Defined WeatherParticleVert in src/opengl/OpenGLTypes.h for type safety
- Updated weather meshes to use the new vertex type
- Integrated GLWeather ownership and usage in MapCanvas
- Updated CMakeLists.txt and removed old display/Weather files
- Ensured all weather shaders continue to use UBOs for performance while aligning with the backend architectural patterns
- Verified build and core tests pass
