+++ src/display/mapcanvas.cpp
@@ -70,16 +70,16 @@ MapCanvas::MapCanvas(MapData &mapData,
     , m_glFont{m_opengl}
     , m_data{mapData}
     , m_groupManager{groupManager}
-    , m_weatherRenderer{std::make_unique<WeatherRenderer>(m_opengl,
-                                                          m_data,
-                                                          m_textures,
-                                                          observer,
-                                                          m_animationManager)}
+    , m_weather{std::make_unique<GLWeather>(m_opengl,
+                                            m_data,
+                                            m_textures,
+                                            m_observer,
+                                            m_animationManager)}
     , m_observer{observer}
 {
     m_animationManager.registerCallback(m_lifetime,
                                         [this]() { return m_batches.remeshCookie.isPending(); });
-    m_weatherRenderer->sig_requestUpdate.connect(m_lifetime, [this]() { update(); });
+    m_weather->sig_requestUpdate.connect(m_lifetime, [this]() { update(); });
 
     m_throttleTimer.setSingleShot(true);
     connect(&m_throttleTimer, &QTimer::timeout, this, [this]() { update(); });
