+++ src/opengl/OpenGL.h
@@ -20,7 +20,7 @@
 
 class FBO;
 class MapCanvas;
-class WeatherRenderer;
+class GLWeather;
 namespace Legacy {
 class Functions;
 } // namespace Legacy
@@ -45,10 +45,7 @@ public:
 
 public:
     NODISCARD const auto &getSharedFunctions(Badge<MapCanvas>) { return getSharedFunctions(); }
-    NODISCARD const auto &getSharedFunctions(Badge<WeatherRenderer>)
-    {
-        return getSharedFunctions();
-    }
+    NODISCARD const auto &getSharedFunctions(Badge<GLWeather>) { return getSharedFunctions(); }
 
 public:
     /* must be called before any other functions */
